smartvenue-ai/services/crowd-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
firebase-admin==6.4.0
google-cloud-pubsub==2.21.1
google-cloud-aiplatform==1.49.0
pydantic==2.6.4
httpx==0.27.0
